Refactor directory structure.
