+++ server/source/serverlib/models/accountmodel.h
@@ -3,8 +3,7 @@
 #include <map>
 
 #include <QSqlQuery>
-
-#include <database.h>
+#include <QString>
 
 class AccountModel
 {
