+++ server/source/serverlib/routes/game.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <route.h>
+#include <routes/route.h>
 
 class Game : public Route
 {
