+++ server/source/serverlib/routes/route.cpp
@@ -2,7 +2,7 @@
 
 #include <QJsonDocument>
 
-#include <database.h>
+#include <utils/database.h>
 
 Route::Route()
 {
