+++ server/tests/server-test/test_utils/mocked_database.h
@@ -1,6 +1,7 @@
 #pragma once
 
-#include <accountmodel.h>
+#include <models/accountmodel.h>
+#include <utils/database.h>
 
 struct ResetDatabase : Database
 {
